+++ .gitignore
@@ -1,2 +1,3 @@
 cut.js
 node_modules
+bundle.js
