+++ index.html
@@ -8,10 +8,6 @@
   </div>
   <div id='console'>hi</div>
   <div id='test-container'></div>
-  <script src='tests.js'></script>
-  <script src='util.js'></script>
-  <script src='calc.js'></script>
-  <script src='math.js'></script>
-  <script src='main.js'></script>
+  <script src='bundle.js'></script>
 </body>
 </html>
