+++ tests.js
@@ -1,0 +1,25 @@
+module.exports = [
+  '123',
+  '1+2',
+  '1+2+',
+  '1/0',
+  '1+2*3/4-5',
+  '5/6+6/7',
+  '1.001+2.56*3.90/0.004-0.5',
+  '2*(3/(1-(5)/6)+4',
+  '2*)(',
+  '2*()',
+  '2*)',
+  '2)(',
+  '2()',
+  '(2)',
+  '()',
+  '(1+2*)',
+  '(1*2)(1+2*)',
+  '(1-2)(1+2)',
+  '(1)(2)',
+  '))((())(',
+  '(((((((())))))))',
+  '()()()()()()()()',
+  '1+2*(3-(1+2-(3/4)+(4/5+(5/6+6/7))/8))'
+]
